refactor: move some generic logic
Move some of the methods in the AnimeFragment into a new SeriesListFragment, these can now be shared
with a MangaFragment
